servicio_decisiones/backend/requirements.txt
flask
flask_sqlalchemy
flask_cors

